pyproject.toml
[tool.poetry]
name = "qnap-img-shrink"
version = "0.1.0"
description = ""
authors = ["Jacek Kotlarski <jacek.kotlarski@bioseco.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jsktoolbox = "^1.2.2"
pillow = "^12.1.0"
pyyaml = "^6.0.3"


[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
pytest = "^9.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
